+++ e2e/e2etests/test_bitcoin_std_deposit_and_call.go
@@ -46,5 +46,10 @@ func TestBitcoinStdMemoDepositAndCall(r *runner.E2ERunner, args []string) {
 	// check if example contract has been called, 'bar' value should be set to amount
 	amountSats, err := zetabitcoin.GetSatoshis(amount)
 	require.NoError(r, err)
-	utils.MustHaveCalledExampleContract(r, contract, big.NewInt(amountSats))
+	utils.MustHaveCalledExampleContract(
+		r,
+		contract,
+		big.NewInt(amountSats),
+		[]byte(r.BTCDeployerAddress.EncodeAddress()),
+	)
 }